fix(Fargate): fix cycle import issue
